refactor: move ShortcutKey model
