+++ src/BookmarkEditor/model.ts
@@ -1,4 +1,5 @@
-import { Bookmark, ShortcutKey } from '../Bookmarks/model'
+import { Bookmark } from '../Bookmarks/model'
+import { ShortcutKey } from '../ShortcutKey/model'
 
 export type EditingBookmark = Bookmark & {
   shortcutKey?: ShortcutKey
