+++ src/Bookmarks/component.tsx
@@ -1,10 +1,11 @@
 import './component.css'
-import { Bookmark, BookmarkFolder, ShortcutMap } from './model'
+import { Bookmark, BookmarkFolder } from './model'
 import React, { FC, ReactElement, useState } from 'react'
 import { useBookmarkFolders, useCollapsedBookmarkFolderIDs, useShortcutMap } from './repository'
 import BookmarkEditorComponent from '../BookmarkEditor/component'
 import { EditingBookmark } from '../BookmarkEditor/model'
 import ShortcutKeyComponent from '../ShortcutKey/component'
+import { ShortcutMap } from '../ShortcutKey/model'
 import { useToggles } from '../Toggles/repository'
 
 const BookmarksComponent: FC = () => {
