+++ src/ShortcutKey/component.ts
@@ -1,4 +1,5 @@
-import { BookmarkFolder, ShortcutMap, shortcutKeyOf } from '../Bookmarks/model'
+import { ShortcutMap, shortcutKeyOf } from './model'
+import { BookmarkFolder } from '../Bookmarks/model'
 import { FC } from 'react'
 import { useGlobalKey } from './infrastructure'
